[•] Add modules meadmin

+++ AyiinXd/modules/__help.py
@@ -1,18 +1,35 @@
-# Copyright (C) 2019 The Raphielscape Company LLC.
+# Copyright (C) 2020 TeamDerUntergang.
 #
-# Licensed under the Raphielscape Public License, Version 1.d (the "License");
-# you may not use this file except in compliance with the License.
+# SedenUserBot is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
 #
-""" Userbot help command """
+# SedenUserBot is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+# @Qulec tarafından yazılmıştır.
+# Thanks @Spechide.
+
+from telethon.errors.rpcerrorlist import BotInlineDisabledError as noinline
+from telethon.errors.rpcerrorlist import YouBlockedUserError
+from telethon.tl.functions.contacts import UnblockRequest
 
 from AyiinXd import CMD_HANDLER as cmd
-from AyiinXd import CMD_HELP, bot, tgbot, ch
+from AyiinXd import CMD_HELP, bot, ch, tgbot
 from AyiinXd.ayiin import ayiin_cmd, eod, eor
 from Stringyins import get_string
 
 
 @ayiin_cmd(pattern="help(?: |$)(.*)")
-async def help(event):
+async def helpyins(event):
+    if event.fwd_from:
+        return
     args = event.pattern_match.group(1).lower()
     if args:
         if args in CMD_HELP:
@@ -20,16 +37,48 @@ async def help(event):
         else:
             await eod(event, get_string("help_10").format(args, cmd))
     else:
-        try:
-            AyiinUBOT = await tgbot.get_me()
-            BOT_USERNAME = AyiinUBOT.username
-            results = await event.client.inline_query(  # pylint:disable=E0602
-                BOT_USERNAME, "@AyiinXdSupport"
+        AyiinUBOT = await tgbot.get_me()
+        BOT_USERNAME = AyiinUBOT.username
+        if BOT_USERNAME is not None:
+            chat = "@Botfather"
+            try:
+                results = await event.client.inline_query(  # pylint:disable=E0602
+                    BOT_USERNAME, "@AyiinXdSupport"
+                )
+                await results[0].click(
+                    event.chat_id, reply_to=event.reply_to_msg_id, hide_via=True
+                )
+                await event.delete()
+            except noinline:
+                xx = await eor(event, "**Inline Mode Tidak aktif.**\n__Sedang Menyalakannya, Harap Tunggu Sebentar...__",
+                               )
+                async with bot.conversation(chat) as conv:
+                    try:
+                        first = await conv.send_message("/setinline")
+                        second = await conv.get_response()
+                        third = await conv.send_message(BOT_USERNAME)
+                        fourth = await conv.get_response()
+                        fifth = await conv.send_message("Search")
+                        sixth = await conv.get_response()
+                        await bot.send_read_acknowledge(conv.chat_id)
+                    except YouBlockedUserError:
+                        await event.client(UnblockRequest(chat))
+                        first = await conv.send_message("/setinline")
+                        second = await conv.get_response()
+                        third = await conv.send_message(BOT_USERNAME)
+                        fourth = await conv.get_response()
+                        fifth = await conv.send_message("Search")
+                        sixth = await conv.get_response()
+                        await bot.send_read_acknowledge(conv.chat_id)
+                    await xx.edit(
+                        f"**Berhasil Menyalakan Mode Inline**\n\n**Ketik** `{cmd}help` **lagi untuk membuka menu bantuan.**"
+                    )
+                await bot.delete_messages(
+                    conv.chat_id,
+                    [first.id, second.id, third.id, fourth.id, fifth.id, sixth.id],
+                )
+        else:
+            await eor(
+                event,
+                "**Silahkan Buat BOT di @BotFather dan Tambahkan Var** `BOT_TOKEN` & `BOT_USERNAME`",
             )
-            await results[0].click(
-                event.chat_id, reply_to=event.reply_to_msg_id, hide_via=True
-            )
-            await event.delete()
-        except Exception as e:
-            await eor(event, get_string("error_1").format(e)
-                      )
